Update import references for refactoring
what is technical debt
